Added documentation to most of Physics
I haven't totaly finished, but I added documentation to most of the base Physics class + the Burgers1D subclass. I still do not know what the Offline subclass thing does, and haven't looked too much into how we compute the Burgers' residual... but I think I got most of the documentation in this commit.

I also made a minor tweak to latent_space.py... the bulk of the changes in this commit are in the physics sub-dictionary, however.